backend/services/genai-service/requirements.txt
google-genai
dotenv
fastapi
uvicorn
spring-config-client
pydantic-settings
py_eureka_client
langchain-google-genai
langgraph
langsmith
langchain[google-genai]
uuid
IPython
langgraph-checkpoint-mongodb
pymongo
langgraph-store-mongodb
pip 
setuptools 
wheel
spacy
nltk
scikit-learn
joblib
PyPDF2
langchain-chroma